refactor(exams): extract features to separate file
